feat: setup prettier and husky pre-commit
