+++ README.md
@@ -1,2 +1,18 @@
-# AccessFrontendExercise
-Interview example, build an website which shows the GitHub users in a List and detail page 
+Access Frontend Exercise
+========================
+
+## Mission
+
+Build an Website which shows the GitHub users in a List
+
+### Technical stack
+
+- Create-react-app
+- Redux
+
+### Todo
+
+- [x] setup Deployment to GitHub page, [document](https://create-react-app.dev/docs/deployment#github-pages)
+- [x] setup ESLint, Prettier, [document](https://create-react-app.dev/docs/setting-up-your-editor#formatting-code-automatically)
+- [ ] List page
+- [ ] Detail Page
